Editing callback and view to rescue error
Errors were generating when the employee id entered was not one that was
represented in the Planned roles.  Now, although the shift is still
created in the callback with employee_id, a message is displyaed to the
user that they need to speak to their manager to get their
assignment. This could be more elegant.