adding nested models for plan entry form
This will allow the user to enter or edit the entire shift plan at once
for the whole shift, putting all employees in planned roles. Then when
the emplyee signs in, their role is announced and the shift is created.
NED to Check if the second part of that works.