adding form for plan sizes
this form is for display only. the data can be submitted, but the model
PlanSizes is not actually functional in the way that it would be in a
working app where the numbers entered would be used to create the
initial labor plan for the next shift.

also some minor cosmetic fixes in this commit